pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "poetry^1.0"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "cognito-emulator"
version = "0.0.0"
description = ""
authors = ["Moriyoshi Koizumi <moriyoshi.koizumi@open-c.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13.8"
authlib = "^0.15"
uvicorn = "^0.11.8"
environs = "^8.0.0"
sqlalchemy = "^1.3.19"
pymysql = "^0.10.1"
ujson = "^3.2.0"
itsdangerous = "^1.1.0"
python-multipart = "^0.0.5"
Jinja2 = "^2.11.2"
aiofiles = "^0.5.0"
argon2-cffi = "^20.1.0"
click = "^7.1.2"
typesystem = "^0.2.4"
srp = "^1.0.16"
flake8 = "^3.8.4"
async_lru = "^1.0.2"
httpx = "^0.16.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.0"
mypy = "^0.782"
black = "^20.8b1"
isort = "^5.5.3"
sqlalchemy-stubs = "^0.3"
alembic = "^1.4.3"
sqlalchemy-seeder = {git = "https://github.com/ricekab/sqlalchemy-seeder.git"}
flake8 = "^3.8.4"

[tool.isort]
profile = "black"
include_trailing_comma = true
multi_line_output = 3

skip = """
    .eggs
    .git
    .mypy_cache
    .nox
    .pytest_cache
    .tox
    .venv
    venv
    alembic
"""
